Update jsd_testing_revised.Rmd
Went through it for grammar, spelling, capitalization, etc. Still needs to be reorganised a bit, and in particular to pay attention to the term 'borrowable' because as I see it, we aren't talking about highly borrowable sounds usually (which we say is a function of frequency) but rather frequently (or infrequently) borrowed sounds, i.e., the empirical frequency in SegBo.